+++ src/include/bench/randomPoints.h
@@ -747,7 +747,7 @@ static std::vector<Rectangle> generateGaiaRectangles() {
 }
 
 static bool is_already_loaded(
-    std::map<std::string, unsigned> &configU,
+    std::map<std::string, uint64_t> &configU,
     Index *spatial_index) {
   if (configU["tree"] == NIR_TREE) {
 
@@ -809,7 +809,7 @@ void repack_tree(T *tree_ptr, std::string &new_file_name,
 }
 
 template <typename T>
-static void runBench(PointGenerator<T> &pointGen, std::map<std::string, unsigned> &configU, std::map<std::string, double> &configD) {
+static void runBench(PointGenerator<T> &pointGen, std::map<std::string, uint64_t> &configU, std::map<std::string, double> &configD) {
   std::cout << "Running benchmark." << std::endl;
 
   // Setup checksums
